trend_coeff now rejects common words

+++ db/schema.rb
@@ -43,13 +43,13 @@ ActiveRecord::Schema.define(version: 20150619153442) do
   add_index "rates", ["user_id"], name: "index_rates_on_user_id", using: :btree
 
   create_table "slogans", force: :cascade do |t|
-    t.string   "body",                       null: false
+    t.string   "body",                         null: false
     t.integer  "rating",         default: 0
     t.integer  "comments_count", default: 0
-    t.integer  "trend_coeff",    default: 0
-    t.integer  "user_id",                    null: false
-    t.datetime "created_at",                 null: false
-    t.datetime "updated_at",                 null: false
+    t.float    "trend_coeff",    default: 0.0
+    t.integer  "user_id",                      null: false
+    t.datetime "created_at",                   null: false
+    t.datetime "updated_at",                   null: false
   end
 
   add_index "slogans", ["user_id"], name: "index_slogans_on_user_id", using: :btree
